+++ features/top_ads/top_ads_management_common/src/main/java/com/tokopedia/topads/common/data/response/GroupInfoResponse.kt
@@ -18,7 +18,7 @@ data class GroupInfoResponse(
                 val groupName: String = "",
 
                 @field:SerializedName("daily_budget")
-                val daiyBudget: Int = 0,
+                val daiyBudget: Float = 0.0F,
 
                 @field:SerializedName("shop_id")
                 val shopId: String = "",
